shift `rules_for_testing` into the test itself
