--- pants-plugins/internal_plugins/rules_for_testing/BUILD
@@ -1,4 +1,0 @@
-# Copyright 2018 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_sources()
